requirements.txt
dify_plugin~=0.0.1b67
aliyun-log-python-sdk>=0.8.0
